requirements.txt
# Core ML and Deep Learning
tensorflow>=2.10.0
numpy>=1.21.0
scikit-learn>=1.0.0
pandas>=1.3.0

# Image Processing
opencv-python>=4.5.0
Pillow>=8.0.0
albumentations>=1.0.0

# Data Visualization
matplotlib>=3.5.0
seaborn>=0.11.0
plotly>=5.0.0

# Web Framework and API
fastapi>=0.68.0
uvicorn>=0.15.0
streamlit>=1.0.0
jinja2>=3.0.0

# Load Testing
locust>=2.0.0

# Utilities
tqdm>=4.60.0
python-multipart>=0.0.5
aiofiles>=0.7.0
requests>=2.25.0

# Jupyter Notebook
jupyter>=1.0.0
ipykernel>=6.0.0 